[Bristol] Add a noscript map using BCC's tiles
This refactors the Map::Zurich module into something that makes it easier
to use custom map tiles for a cobrand. Updating Zurich to use this new
WMTSBase class will be a future commit.

This also includes some refactoring of the JS & map templates
to make it easier to use a custom WMTS base map in future
cobrands.